4_Flask/Flask_App/requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
azure.cosmos==4.0.0b6
azure-core==1.5.0
azure-common==1.1.23
pandas